Add go.mod file for tracking dependencies
The method call `RoundTrip` does not allow access to the Header section of the response. The passed in `out` body only captures the contents of the body section. Setting this to allow the caller to capture both the header and body of the SOAP response. A more structure caller interface on the pipeline as I plan for a Merge Request on the original repo.